Rename incorrectly named object.
